+++ nextjs-logs-manager/app/shared/types.ts
