+++ nextjs-logs-manager/app/shared/utils.ts
